pyproject.toml
[tool.poetry]
name = "money-man"
version = "0.1.0"
description = ""
authors = ["Udi Oron <udioron@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.6"
pytest-django = "^4.8.0"
ipython = "^8.24.0"
faker = "^25.1.0"
django-extensions = "^3.2.3"
tqdm = "^4.66.4"
django-bootstrap5 = "^24.2"
django-flatpickr = "^2.0.3"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "money_man.settings"
